test: separate MCP wire coverage
Keep registry checks in the unit lane and run full MCP transport verification as integration coverage to avoid repeated server construction.
