+++ test/server/tools/daemonModeTools.test.ts
@@ -1,10 +1,8 @@
-import { afterAll, afterEach, beforeAll, beforeEach, describe, expect, test } from "bun:test";
-import { z } from "zod/v4";
-import { createMcpServer } from "../../../src/server/index";
+import { afterEach, beforeEach, describe, expect, test } from "bun:test";
+import { registerMcpTools } from "../../../src/server/index";
 import { ToolRegistry } from "../../../src/server/toolRegistry";
 import { serverConfig } from "../../../src/utils/ServerConfig";
 import { setDebugModeEnabled } from "../../../src/utils/debug";
-import { McpTestFixture } from "../../fixtures/mcpTestFixture";
 
 describe("Daemon-only MCP tools", () => {
   beforeEach(() => {
@@ -18,7 +16,7 @@ describe("Daemon-only MCP tools", () => {
   });
 
   test("registers plan tools in both modes, criticalSection only in daemon mode", () => {
-    createMcpServer();
+    registerMcpTools(false);
 
     const toolNames = ToolRegistry.getToolDefinitions().map((tool) => tool.name);
     expect(toolNames).toContain("executePlan");
@@ -26,7 +24,7 @@ describe("Daemon-only MCP tools", () => {
   });
 
   test("registers criticalSection/barrier plan-only in daemon mode (hidden from discovery, usable in plans)", () => {
-    createMcpServer({ daemonMode: true });
+    registerMcpTools(true);
 
     const toolNames = ToolRegistry.getToolDefinitions().map((tool) => tool.name);
     expect(toolNames).toContain("executePlan");
@@ -42,7 +40,7 @@ describe("Daemon-only MCP tools", () => {
   });
 
   test("hides debug-only tools unless debug mode is enabled", () => {
-    createMcpServer();
+    registerMcpTools(false);
 
     expect(ToolRegistry.getTool("accessibilityFocus")).toBeUndefined();
     expect(ToolRegistry.getTool("setUIState")).toBeUndefined();
@@ -55,7 +53,7 @@ describe("Daemon-only MCP tools", () => {
 
     (ToolRegistry as any).tools.clear();
     setDebugModeEnabled(true);
-    createMcpServer({ debug: true });
+    registerMcpTools(false);
 
     expect(ToolRegistry.getTool("accessibilityFocus")).toBeDefined();
     expect(ToolRegistry.getTool("setUIState")).toBeDefined();
@@ -79,7 +77,7 @@ describe("Daemon-only MCP tools", () => {
       "getNetworkGraph",
     ];
 
-    createMcpServer();
+    registerMcpTools(false);
 
     const defaultNames = ToolRegistry.getToolDefinitions().map((tool) => tool.name);
     for (const toolName of embeddedSdkTools) {
@@ -89,67 +87,12 @@ describe("Daemon-only MCP tools", () => {
 
     (ToolRegistry as any).tools.clear();
     serverConfig.setEmbeddedSdkEnabled(true);
-    createMcpServer();
+    registerMcpTools(false);
 
     const embeddedSdkNames = ToolRegistry.getToolDefinitions().map((tool) => tool.name);
     for (const toolName of embeddedSdkTools) {
       expect(ToolRegistry.getTool(toolName)).toBeDefined();
       expect(embeddedSdkNames).toContain(toolName);
     }
-  });
-});
-
-// Issue #4181, rank 1 (A1): the registry-level assertions above prove
-// getToolDefinitions()/getTool() hide these tools, but the regression that
-// matters is over the WIRE — a raw MCP client must not be able to call them.
-// The CallTool handler resolves names through getTool() (index.ts). Mutating
-// that resolve to getToolForPlan() would make criticalSection/barrier/setUIState
-// directly callable; these rows red on that mutation because the call would no
-// longer be rejected as an unknown tool.
-describe("plan-only/debug-only tools are ungated over the wire", () => {
-  let fixture: McpTestFixture;
-
-  const permissiveResult = z.object({}).passthrough();
-
-  beforeAll(async () => {
-    (ToolRegistry as any).tools.clear();
-    // Daemon mode registers criticalSection/barrier plan-only; debug OFF keeps
-    // setUIState plan-only (resolvable via getToolForPlan, hidden from getTool).
-    fixture = new McpTestFixture({ daemonMode: true });
-    await fixture.setup();
-  });
-
-  afterAll(async () => {
-    if (fixture) {
-      await fixture.teardown();
-    }
-    (ToolRegistry as any).tools.clear();
-  });
-
-  test.each(["criticalSection", "barrier", "setUIState"])(
-    "%s is rejected as an unknown tool when called directly over the wire",
-    async (name) => {
-      const { client } = fixture.getContext();
-      await expect(
-        client.request({ method: "tools/call", params: { name, arguments: {} } }, permissiveResult),
-      ).rejects.toThrow("Unknown tool");
-    },
-  );
-
-  test("a normally-advertised tool is NOT rejected as unknown (control)", async () => {
-    // executePlan is advertised in both modes, so it resolves through getTool
-    // and does not produce an "Unknown tool" rejection (it may fail validation,
-    // but not with that message).
-    const { client } = fixture.getContext();
-    let message = "";
-    try {
-      await client.request(
-        { method: "tools/call", params: { name: "executePlan", arguments: {} } },
-        permissiveResult,
-      );
-    } catch (error) {
-      message = error instanceof Error ? error.message : String(error);
-    }
-    expect(message).not.toContain("Unknown tool");
   });
 });
